#350 findbugs fixes

+++ back-end/src/test/java/ee/hm/dop/rest/content/LearningObjectResourceTest.java
@@ -199,13 +199,13 @@ public class LearningObjectResourceTest extends ResourceIntegrationTestBase {
     @Test
     public void increaseViewCountNotExistingMaterial() {
         login(USER_ADMIN);
-        Response response = doGet(format(GET_MATERIAL_URL, (long) NOT_EXISTS_ID));
+        Response response = doGet(format(GET_MATERIAL_URL, NOT_EXISTS_ID));
         assertEquals(Status.NO_CONTENT.getStatusCode(), response.getStatus());
 
         response = doPost(INCREASE_VIEW_COUNT_URL, materialWithId(NOT_EXISTS_ID));
         assertEquals(Status.INTERNAL_SERVER_ERROR.getStatusCode(), response.getStatus());
 
-        response = doGet(format(GET_MATERIAL_URL, (long) NOT_EXISTS_ID));
+        response = doGet(format(GET_MATERIAL_URL, NOT_EXISTS_ID));
         assertEquals(Status.NO_CONTENT.getStatusCode(), response.getStatus());
     }
